# docs/openapi.yml
# THIS FILE IS GENERATED BY OpenApiSchemaExportTest — DO NOT EDIT MANUALLY.
# Run `./mvnw -pl service -am test -Dtest=OpenApiSchemaExportTest -Dsurefire.failIfNoSpecifiedTests=false` to regenerate.
---
openapi: 3.1.0
components:
  schemas:
    CreateEventRequest:
      type: object
      required:
      - name
      properties:
        name:
          type: string
          pattern: \S
        occurredAt:
          $ref: "#/components/schemas/Instant"
        payload:
          type: string
        description:
          type: string
        parentId:
          type: string
        environment:
          type: string
    EventDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        occurredAt:
          $ref: "#/components/schemas/Instant"
        payload:
          type: string
        description:
          type: string
        parentId:
          type: string
        environment:
          type: string
        createdAt:
          $ref: "#/components/schemas/Instant"
        updatedAt:
          $ref: "#/components/schemas/Instant"
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
    PublishEventRequest:
      type: object
      required:
      - name
      - occurredAt
      properties:
        name:
          type: string
          pattern: \S
        occurredAt:
          $ref: "#/components/schemas/Instant"
        payload:
          type: string
        description:
          type: string
        parentId:
          type: string
        environment:
          type: string
    Response:
      type: object
      properties:
        event:
          $ref: "#/components/schemas/EventDto"
    Response1:
      type: object
      properties:
        names:
          type: array
          items:
            type: string
    Response2:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/EventDto"
        nextCursor:
          type: string
    Response3:
      type: object
      properties:
        event:
          $ref: "#/components/schemas/EventDto"
    Response4:
      type: object
      properties:
        success:
          type: boolean
    Response5:
      type: object
      properties:
        event:
          $ref: "#/components/schemas/EventDto"
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: http://dev-qits-platform-idp:8080/idp/.well-known/openid-configuration
      description: Authentication
paths:
  /events/api/events:
    get:
      summary: List
      operationId: listEvents
      parameters:
      - name: attr
        in: query
        schema:
          type: array
          items:
            type: string
      - name: cursor
        in: query
        schema:
          type: string
      - name: environment
        in: query
        schema:
          type: string
      - name: limit
        in: query
        schema:
          type: string
      - name: name
        in: query
        schema:
          type: string
      - name: order
        in: query
        schema:
          type: string
      - name: parentId
        in: query
        schema:
          type: string
      - name: q
        in: query
        schema:
          type: string
      - name: since
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response2"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Event Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:system
        - qits:agent
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response5"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Create
      tags:
      - Event Controller
      security:
      - SecurityScheme:
        - qits:admin
  /events/api/events/names:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response1"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Names
      tags:
      - Event Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
  /events/api/events/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishEventRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Publish
      tags:
      - Event Controller
      security:
      - SecurityScheme:
        - qits:system
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response3"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get
      tags:
      - Event Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response4"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Delete
      tags:
      - Event Controller
      security:
      - SecurityScheme:
        - qits:admin
  /events/api/stream:
    get:
      summary: Live event stream (Server-Sent Events)
      description: "Pushes each newly created event whose name is named by ?names=\
        \ (comma-separated; '*' means everything; absent means nothing) as an SSE\
        \ event whose id: field is the event id and whose data is the EventCreated\
        \ envelope. Live only: no replay, no offset, no catch-up."
      parameters:
      - name: names
        in: query
        schema:
          type: string
      responses:
        "200":
          description: "An open SSE stream. Each event carries the event id in its\
            \ id: field and the EventCreated envelope's JSON in its data: field; comment\
            \ lines mark the open and keep the connection alive. The body never ends."
          content:
            text/event-stream:
              schema:
                type: string
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Event Stream Resource
      security:
      - SecurityScheme:
        - qits:admin
        - qits:system
        - qits:agent
info:
  title: qits-events API
  version: 2026.1002.180127